+++ Scripts/Runtime/TMP_Settings.cs
@@ -18,7 +18,7 @@ namespace TMPro
         /// </summary>
         public static string version
         {
-            get { return "1.2.3"; }
+            get { return "1.3.0"; }
         }
 
         /// <summary>