+++ Scripts/Runtime/TMPro_Private.cs
@@ -1,5 +1,5 @@
-//#define PROFILE_ON
-//#define PROFILE_PHASES_ON
+//#define TMP_PROFILE_ON
+//#define TMP_PROFILE_PHASES_ON
 
 using UnityEngine;
 using System;
@@ -2400,6 +2400,9 @@ namespace TMPro
                         m_lastVisibleCharacterOfLine = m_characterCount;
                         m_textInfo.lineInfo[m_lineNumber].spaceCount += 1;
                         m_textInfo.spaceCount += 1;
+
+                        if (charCode == 0xA0)
+                            m_textInfo.lineInfo[m_lineNumber].controlCharacterCount += 1;
                     }
                     else
                     {
@@ -2439,8 +2442,6 @@ namespace TMPro
                     {
                         m_textInfo.lineInfo[m_lineNumber].spaceCount += 1;
                         m_textInfo.spaceCount += 1;
-
-                        if (charCode == 0xA0) m_textInfo.lineInfo[m_lineNumber].controlCharacterCount = +1;
                     }
                 }
                 #endregion Handle Visible Characters